fix: compile wasm test actors separately
Compile each test actor as _either_ a byte blob containing the wasm
module, or as a wasm module.

Previously, we'd compile all test actors into the _test_ binary. But
really, we only want them compiled into individual wasm binaries. This
sort of worked for some reason, but really shouldn't have, as each wasm
module exported a non-mangled "invoke" function.

Unfortunately, it looks like it stopped "sort of working" the second
@mriise added a _second_ exported function.
