+++ testing/integration/tests/fil-hello-world-actor/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-[dependencies]
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 fvm_sdk = { version = "3.0.0-alpha.2", path = "../../../../sdk" }
 fvm_shared = { version = "3.0.0-alpha.2", path = "../../../../shared" }
