+++ testing/integration/tests/fil-malformed-syscall-actor/Cargo.toml
@@ -4,9 +4,9 @@ version = "0.1.0"
 edition = "2021"
 publish = false
 
-[dependencies]
-fvm_sdk = { version = "3.0.0-alpha.2", path = "../../../../sdk" }
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 fvm_shared = { version = "3.0.0-alpha.2", path = "../../../../shared" }
+fvm_sdk = { version = "3.0.0-alpha.2", path = "../../../../sdk" }
 
 [build-dependencies]
 substrate-wasm-builder = "4.0.0"
