+++ testing/integration/tests/fil-malformed-syscall-actor/src/lib.rs
@@ -1,9 +1,11 @@
-use fvm_sdk::sys::crypto::compute_unsealed_sector_cid;
-
+#[cfg(not(target_arch = "wasm32"))]
 include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
 
 #[no_mangle]
+#[cfg(target_arch = "wasm32")]
 pub fn invoke(_: u32) -> u32 {
+    use fvm_sdk::sys::crypto::compute_unsealed_sector_cid;
+
     let piece: Vec<u8> = vec![];
     let mut cid: Vec<u8> = vec![];
